+++ src/components/Title/Title.js
@@ -23,6 +23,7 @@ const Title = ({
 Title.displayName = 'Title'
 
 Title.propTypes = {
+  as: PropTypes.node,
   className: PropTypes.string,
   spaced: PropTypes.bool,
   is: PropTypes.oneOf(['1', '2', '3', '4', '5', '6'])