+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0
+
+* Publish to NPM
+* Add ESLint
+
+## 0.0.7
+
+* Add missing props from Input
+* Add Input, Link to styleguide
+
 ## 0.0.6
 
 * Fix button warning
@@ -15,8 +25,8 @@
 
 ## 0.0.3
 
-* fix class names
-* fix styleguide Message
+* Fix class names
+* Fix styleguide Message
 * Add Control, Field and Haeding to styleguide
 
 ## 0.0.2